bert-classifier-server/requirements.txt
tensorflow==2.4.1
Flask==2.1.2
numpy==1.19.5
tf-models-official==2.3.0
verboselogs==1.7
coloredlogs==15.0
